add bq, snowflake tests for seed type overrides
