+++ test/integration/005_simple_seed_test/models-pg/schema.yml
@@ -1,5 +1,4 @@
 
-
 seed_enabled:
     constraints:
         column_type:
